requirements.txt
fastapi==0.70.0
mccabe==0.6.1
pydantic==1.8.2
pyflakes==2.4.0
pymongo[tls,srv,gssapi]==4.0.1
pytest==6.2.5
pytest-cov==3.0.0
PyYAML==6.0
requests==2.26.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.2
